Move SchemaTree to LyvEffectiveModelContextFactory
Move initialization SchemaTree to LyvEffectiveModelContextFactory
so it will not be executed for every model individually.
